refactor: move color file to helpers
